+++ runtime/databricks/automl_runtime/sklearn/preprocessing/__init__.py
@@ -14,4 +14,7 @@
 # limitations under the License.
 #
 
-__all__ = ["compose", "feature_selection", "impute", "preprocessing"]
+from .timestamp_transformer import TimestampTransformer
+from .date_transformer import DateTransformer
+
+__all__ = ["DateTransformer", "TimestampTransformer"]
