+++ runtime/databricks/automl_runtime/sklearn/preprocessing/timestamp_transformer.py
@@ -16,29 +16,33 @@
 
 import pandas as pd
 
-from databricks.automl_runtime.sklearn.base_datetime_transformer import BaseDatetimeTransformer
+from databricks.automl_runtime.sklearn.preprocessing.base_datetime_transformer import BaseDatetimeTransformer
 
 
-class DateTransformer(BaseDatetimeTransformer):
+class TimestampTransformer(BaseDatetimeTransformer):
     """
-    Generate features from date column.
+    Generate features from timestamp or datetime-like string column.
     """
+
+    HOUR_COLUMN_INDEX = 10
 
     def transform(self, X):
         """
-        Transform date data to datetime features.
+        Transform timestamp data to datetime features.
 
         Parameters
         ----------
         X : pd.DataFrame of shape (n_samples, 1)
-            The only column is a date column.
+            The only column is either a timestamp column or string column with
+            datetime values encoded in ISO 8601 format.
 
         Returns
         -------
-        X_tr : pd.DataFrame of shape (n_samples, 10)
+        X_tr : pd.DataFrame of shape (n_samples, 13)
             Transformed features.
         """
-        X.iloc[:, 0] = X.iloc[:, 0].apply(pd.to_datetime, errors="coerce")
+        # Convert column to datetime if data type is string and standardize to UTC
+        X.iloc[:, 0] = X.iloc[:, 0].apply(pd.to_datetime, errors="coerce", utc=True).dt.tz_localize(None)
         X = X.fillna(pd.Timestamp(self.EPOCH))  # Fill NaT with the Unix epoch
 
-        return self._generate_datetime_features(X, include_timestamp=False)
+        return self._generate_datetime_features(X, include_timestamp=True)
