+++ runtime/tests/automl_runtime/sklearn/preprocessing/datetime_transformer_e2e_test.py
@@ -23,8 +23,7 @@ from sklearn.compose import ColumnTransformer
 from sklearn.pipeline import Pipeline
 from sklearn.preprocessing import OneHotEncoder
 
-from databricks.automl_runtime.sklearn import DateTransformer
-from databricks.automl_runtime.sklearn import TimestampTransformer
+from databricks.automl_runtime.sklearn.preprocessing import DateTransformer, TimestampTransformer
 
 
 class TestDatetimeTransformer(unittest.TestCase):
