+++ runtime/tests/automl_runtime/sklearn/preprocessing/timestamp_transformer_test.py
@@ -20,7 +20,7 @@ import pandas as pd
 import numpy as np
 from sklearn.pipeline import Pipeline
 
-from databricks.automl_runtime.sklearn import TimestampTransformer
+from databricks.automl_runtime.sklearn.preprocessing import TimestampTransformer
 
 
 class TestTimestampTransformer(unittest.TestCase):
